clean-up bugs and address codebase changes

+++ components/iota_ui_frames.py
@@ -5,7 +5,7 @@ from six.moves import range
 """
 Author      : Lyubimov, A.Y.
 Created     : 01/17/2017
-Last Changed: 08/01/2019
+Last Changed: 09/18/2019
 Description : IOTA GUI Windows / frames
 """
 
@@ -354,7 +354,6 @@ class MainWindow(IOTABaseFrame):
                 lambda evt: self.open_options_dialog(
                     phil_index=self.bknd_index,
                     include=[
-                        "verbosity",
                         "geometry",
                         "profile",
                         "prediction",
@@ -678,7 +677,15 @@ class MainWindow(IOTABaseFrame):
                 import shutil
 
                 try:
-                    shutil.rmtree(self.proc_window.info.tmp_base)
+                    # Clean up temp folders and files if necessary
+                    if os.path.isdir(self.proc_window.info.tmp_base):
+                        shutil.rmtree(self.proc_window.info.tmp_base)
+
+                    dials_log_dir = os.path.join(
+                        self.proc_window.info.log_base, "logs/dials_logs"
+                    )
+                    if os.path.isdir(dials_log_dir):
+                        shutil.rmtree(dials_log_dir)
                 except Exception:
                     pass
                 print("JOB TERMINATED!")
@@ -2707,7 +2714,9 @@ class ProcWindow(IOTABaseFrame):
 
             try:
                 shutil.rmtree(self.info.tmp_base)
-            except Exception:
+                shutil.rmtree(os.path.join(self.info.log_base, "dials_logs"))
+            except Exception as e:
+                print("DEBUG: UNABLE TO DELETE FOLDER: {}".format(e))
                 pass
 
         # Export final info file
